feat: Add posts relationship to User model
This commit adds a new method `posts()` to the User model, which defines a `hasMany` relationship with the Post model. This relationship will be used to retrieve all the posts associated with a user.

Recent repository commits follow the convention of starting with a chore prefix and providing a brief description of the changes made. This commit message follows the same convention, starting with a feat prefix to indicate a new feature and providing a clear description of the changes made.