refactor: Remove create new post button from author view
This commit removes the "Create New Post" button from the author view in the `author.blade.php` file. The button was previously located in the `div` with the class `flex justify-end gap-4`. This change aligns with recent updates to the PostController and improves the user experience by removing the option to create a new post from the author view. The commit message follows the established convention of starting with a refactor prefix and provides a clear description of the changes made.