refactor: Add 'username' field to User model
This commit adds the 'username' field to the User model in the app/Models/User.php file. The 'username' field is included in the $fillable array, allowing it to be mass assignable. This change enables the application to store and retrieve the username information for each user. The commit message follows the established convention of starting with a refactor prefix and provides a clear description of the changes made.